refactor: MemberService 패키지 이동
